pip-requirements.txt
djangorestframework
dj-rest-auth
django-allauth[socialaccount]